[Feat] 출력 로직 콘솔로 이동
