+++ src/main/java/com/prgrms/voucher_manage/domain/customer/controller/CustomerController.java
@@ -1,15 +1,18 @@
 package com.prgrms.voucher_manage.domain.customer.controller;
 
+import com.prgrms.voucher_manage.domain.customer.entity.Customer;
 import com.prgrms.voucher_manage.domain.customer.service.CustomerService;
 import lombok.RequiredArgsConstructor;
 import org.springframework.stereotype.Controller;
+
+import java.util.List;
 
 @Controller
 @RequiredArgsConstructor
 public class CustomerController {
     private final CustomerService customerService;
 
-    public void getBlackList() {
-        customerService.getBlackList();
+    public List<Customer> getBlackList() {
+        return customerService.getBlackList();
     }
 }
