+++ src/main/java/com/prgrms/voucher_manage/domain/customer/service/CustomerService.java
@@ -2,7 +2,6 @@ package com.prgrms.voucher_manage.domain.customer.service;
 
 import com.prgrms.voucher_manage.domain.customer.entity.Customer;
 import com.prgrms.voucher_manage.domain.customer.repository.CustomerRepository;
-import com.prgrms.voucher_manage.console.OutputUtil;
 import lombok.RequiredArgsConstructor;
 import org.springframework.stereotype.Service;
 
@@ -11,11 +10,9 @@ import java.util.List;
 @Service
 @RequiredArgsConstructor
 public class CustomerService {
-    private final OutputUtil outputUtil;
     private final CustomerRepository customerRepository;
 
-    public void getBlackList() {
-        List<Customer> customers = customerRepository.findAll();
-        customers.forEach(outputUtil::printCustomerInfo);
+    public List<Customer> getBlackList() {
+        return customerRepository.findAll();
     }
 }
