+++ src/main/java/com/prgrms/voucher_manage/domain/voucher/controller/VoucherController.java
@@ -1,11 +1,14 @@
 package com.prgrms.voucher_manage.domain.voucher.controller;
 
+import com.prgrms.voucher_manage.domain.voucher.entity.Voucher;
 import com.prgrms.voucher_manage.domain.voucher.entity.VoucherType;
 import com.prgrms.voucher_manage.domain.voucher.dto.CreateVoucherDto;
 import com.prgrms.voucher_manage.domain.voucher.service.VoucherService;
 import com.prgrms.voucher_manage.exception.InvalidDiscountRangeException;
 import lombok.RequiredArgsConstructor;
 import org.springframework.stereotype.Controller;
+
+import java.util.List;
 
 import static com.prgrms.voucher_manage.domain.voucher.entity.VoucherType.FIXED;
 import static com.prgrms.voucher_manage.domain.voucher.entity.VoucherType.PERCENT;
@@ -27,8 +30,8 @@ public class VoucherController {
         voucherService.createVoucher(new CreateVoucherDto(voucherType, discountAmount));
     }
 
-    public void getVouchers() {
-        voucherService.getVouchers();
+    public List<Voucher> getVouchers() {
+        return voucherService.getVouchers();
     }
 
     public boolean isInvalidPercent(Long discountAmount) {
